// src/app/interfaces/Trabalho.ts
export interface Trabalho {
    avaliador: string;
    idAvaliador: string;
    idTrabalho: string;
    titulo: string;
    apresentador: string;
    orientador: string;
    evento: string;
    dia: string;
    horario: string;
    predio: string;
    sala: string;
    painel: string;
    modulo: string;
    idModulo: string;
    tipoForm: string;
}