// src/app/interfaces/Avaliacao.ts
export interface Avaliacao {
    id: string;
    nomeTrabalho: string;
    avaliadorOriginal: string;
    avaliadorReal: string;
    q1: string;
    q2: string;
    q3: string;
    q4: string;
    q5: string;
    q6: string;
    q7: string;
    q8: string;
    q9: string;
    q10: string;
    estado: string;
    timestamp: string;
}